+++ configs/solutions_team/jax/solutionsteam_flax_latest_supported_config.py
