+++ utils/metric.py
@@ -30,8 +30,7 @@ from apis import gcp_config, test_config
 from apis import metric_config
 from configs import composer_env
 from google.cloud import storage
-from implementations.utils import bigquery
-from implementations.utils import composer
+from utils import bigquery, composer
 import jsonlines
 import numpy as np
 import tensorflow as tf
